+++ src/main/scala/com/databricks/spark/sql/perf/mllib/classification/LogisticRegression.scala
@@ -1,96 +1,14 @@
 package com.databricks.spark.sql.perf.mllib.classification
 
-import org.apache.spark.ml.ModelBuilder
-import org.apache.spark.ml.classification.{LogisticRegression, LogisticRegressionModel}
-import org.apache.spark.ml.evaluation.{Evaluator, MulticlassClassificationEvaluator}
-import org.apache.spark.ml.linalg.Vectors
-import org.apache.spark.sql.{DataFrame, Row}
-import org.apache.spark.sql.functions._
-
-import com.databricks.spark.sql.perf.mllib.{ClassificationContext, ClassificationPipelineDescription}
 import com.databricks.spark.sql.perf.mllib.OptionImplicits._
+import com.databricks.spark.sql.perf.mllib._
 import com.databricks.spark.sql.perf.mllib.data.DataGenerator
 
-trait ScoringWithEvaluator {
-  self: ClassificationPipelineDescription =>
-
-  protected def evaluator(ctx: ClassificationContext): Evaluator
-
-  final override def score(
-      ctx: ClassificationContext,
-      testSet: DataFrame,
-      model: Model): Double = {
-    val eval = model.transform(testSet)
-    evaluator(ctx).evaluate(eval)
-  }
-}
-
-trait TrainingSetFromTransformer {
-  self: ClassificationPipelineDescription =>
-
-  protected def initialData(ctx: ClassificationContext): DataFrame
-
-  protected def initialModel(ctx: ClassificationContext): Model
-
-  final override def trainingDataSet(ctx: ClassificationContext): DataFrame = {
-    val initial = initialData(ctx)
-    val model = initialModel(ctx)
-    model.transform(initial).select(col("features"), col("prediction").as("label"))
-  }
-}
-
-trait TestFromTraining {
-  self: ClassificationPipelineDescription =>
-
-  final override def testDataSet(ctx: ClassificationContext): DataFrame = {
-    // Copy the context with a new seed.
-    val ctx2 = ctx.commonParams.randomSeed match {
-      case Some(x) =>
-        val p = ctx.commonParams.copy(randomSeed = Some(x + 1))
-        ctx.copy(commonParams = p)
-      case None =>
-        // Making a full copy to reset the internal seed.
-        ctx.copy()
-    }
-    self.trainingDataSet(ctx2)
-  }
-}
-
-object LogisticRegressionBenchmark
-  extends ClassificationPipelineDescription {
-
-  override type Model = LogisticRegressionModel
-
-  def trainingDataSet(ctx: ClassificationContext): DataFrame = {
-    DataGenerator.generateBinaryLabeledPoints(
-      ctx.sqlContext,
-      ctx.commonParams.numExamples.get,
-      ctx.commonParams)
-  }
-
-  def testDataSet(ctx: ClassificationContext): DataFrame = {
-    DataGenerator.generateBinaryLabeledPoints(
-      ctx.sqlContext,
-      ctx.commonParams.numTestExamples.get,
-      ctx.commonParams)
-  }
-
-  def train(ctx: ClassificationContext,
-            trainingSet: DataFrame): Model = {
-    import ctx.extraParams._
-    val lr = new LogisticRegression()
-        .setTol(tol)
-        .setRegParam(regParam)
-    lr.fit(trainingSet)
-  }
-
-  def score(ctx: ClassificationContext,
-            testSet: DataFrame,
-            model: Model): Double = {
-    val eval = model.evaluate(testSet)
-    Utils.accuracy(eval.predictions)
-  }
-}
+import org.apache.spark.ml.ModelBuilder
+import org.apache.spark.ml.classification.{LogisticRegression, LogisticRegressionModel}
+import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
+import org.apache.spark.ml.linalg.Vectors
+import org.apache.spark.sql.{DataFrame, Row}
 
 object LogisticRegressionBenchmark2 extends ClassificationPipelineDescription
   with TestFromTraining with TrainingSetFromTransformer with ScoringWithEvaluator {
